Move http_auth.rs under crates/types/src/schema/deployment/
Per Francesco's review nudge, the persisted HttpAuth and GoogleIdTokenAuth
types live alongside the rest of the persisted deployment record under
schema/deployment/, rather than in the top-level deployment module that
also carries the deployment-address surface. The derive_audience helper
moves with them since the schema registry materialisation pass and the
service-client dispatch fallback both consume it from this same module.

The restate_types::deployment::{HttpAuth, GoogleIdTokenAuth,
derive_audience} re-export is preserved so external consumers keep
working unchanged.
