+++ crates/types/src/schema/deployment/http_auth.rs
